backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=2.0
djangorestframework-camel-case>=1.4
django-cors-headers>=4.3
django-filter>=24.0
django-storages>=1.14
django-ratelimit>=4.1
drf-spectacular>=0.29
celery>=5.4
redis>=5.0
boto3>=1.34
psycopg2-binary>=2.9
Pillow>=10.0
factory-boy>=3.3
pytest-django>=4.8
gunicorn>=22.0
